requirements.txt
fastapi[standard]
sqlalchemy
psycopg[binary]
redis
python-dotenv
pydantic-settings
faststream[redis]
elasticsearch[async]
pymongo
beanie
aiosqlite
